backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
networkx==3.2.1
firebase-admin==6.2.0
google-genai==0.3.0
python-dotenv==1.0.0
python-multipart==0.0.6
openpyxl==3.1.2
pydantic==2.5.2
